Moved accessibility enums to seperate folder and updated imports
