+++ src/atoms/buttons/button.tsx
@@ -3,7 +3,7 @@ import { ButtonStyles } from "../constants/button-styles";
 import { ButtonTypes } from "../constants/button-types";
 import React, { forwardRef } from "react";
 import { StringUtils } from "andculturecode-javascript-core";
-import { AccessibilityLabels } from "../constants/accessibility-labels";
+import { AccessibilityLabels } from "../../enums/accessibility-labels";
 
 // -----------------------------------------------------------------------------------------
 // #region Interfaces
